load_env_vars.py
# load_env_vars.py

from dotenv import load_dotenv
import os

load_dotenv()

# Access the environment variables
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
GITHUB_TOKEN = os.getenv("GITHUB_TOKEN")


