requirements.txt
PyGithub
openai
pyyaml
python-dotenv
